Move more code to prep from resolveNestedMatch
